requirements.txt
biopython
nose
rednose
